requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
python-decouple
psycopg2-binary
email-validator
passlib[bcrypt]
PyJWT
